requirements.txt
appdirs==1.4.4
asgiref==3.5.2
atomicwrites==1.4.0
attrs==21.4.0
beautifulsoup4==4.11.1
black==22.3.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
Django==4.0.4
django-js-asset==2.0.0
django-mptt==0.13.4
django-treebeard==4.5.1
factory-boy==3.2.1
Faker==13.12.0
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
Mako==1.2.0
MarkupSafe==2.1.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==6.2.5
pytest-base-url==1.4.2
pytest-django==4.5.2
pytest-factoryboy==2.3.1
pytest-html==3.1.1
pytest-metadata==2.0.1
pytest-selenium==3.0.0
pytest-variables==1.9.0
python-dateutil==2.8.2
python-dotenv==0.20.0
requests==2.27.1
selenium==3.141.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tenacity==6.3.1
toml==0.10.2
tomli==2.0.1
typing_extensions==4.2.0
tzdata==2022.1
urllib3==1.26.9
webdriver-manager==3.7.0
